// src/app/models/lecture-info.ts
export interface ILectureInfo {
  android: any;
  creationTimestamp: number;
  documentId?: string;
  documentPath: string;
  downloadPlace: number;
  favouritePlace: number;
  id: number;
  ios: any;
  isCompleted: boolean;
  isDownloaded: boolean;
  isFavourite: boolean;
  isInPrivateList: boolean;
  isInPublicList: boolean;
  lastModifiedTimestamp: number;
  lastPlayedPoint: number;
  privateListIDs: any[];
  publicListIDs: any[];
  totalPlayedNo: number;
  totalPlayedTime: number;
  totallength: number;
}